Added solid error handling
