+++ services/web/server/src/simcore_service_webserver/studies_dispatcher/__init__.py
@@ -1,0 +1,5 @@
+"""
+    This app module dispatches pre-configure or on-the-fly studies to user from a link request
+    that is redirected to the front-end which customizes its view
+"""
+# TODO: move here all studies_access.py
