+++ services/web/server/src/simcore_service_webserver/studies_dispatcher/_core.py
@@ -1,10 +1,17 @@
 from dataclasses import dataclass
-from typing import Optional
 import uuid
-
+from aiohttp import web
+from pydantic import ValidationError
 from functools import lru_cache
 
 BASE_UUID = uuid.UUID("ca2144da-eabb-4daf-a1df-a3682050e25f")
+
+
+# VIEWERS MAP -----------------------------------------------------------------------------
+class MatchNotFoundError(Exception):
+    def __init__(self, reason):
+        super().__init__()
+        self.reason = reason
 
 
 @dataclass
@@ -21,13 +28,6 @@ class ViewerInfo:
 SIM4LIFE_VIEWER = ViewerInfo("simcore/services/dynamic/sim4life", "1.0.16", "sim4life")
 
 
-
-class MatchNotFoundError(Exception):
-    def __init__(self, reason):
-        super().__init__()
-        self.reason = reason
-
-
 def find_compatible_viewer(file_size: int, file_type: str) -> ViewerInfo:
     # FIXME: temporarily hard-coded
     # if file_size>MAXLIMIT:
@@ -39,9 +39,22 @@ def find_compatible_viewer(file_size: int, file_type: str) -> ViewerInfo:
     raise MatchNotFoundError("No")
 
 
+# UTILITIES ---------------------------------------------------------------
+
 
 @lru_cache()
-def compose_uuid_from( *values ) -> str:
+def compose_uuid_from(*values) -> str:
     composition = "/".join(map(str, values))
     new_uuid = uuid.uuid5(BASE_UUID, composition)
     return str(new_uuid)
+
+
+class ValidationMixin:
+    @classmethod
+    def create_from(cls, request: web.Request):
+        try:
+            obj = cls(**request.query.keys())
+        except ValidationError as err:
+            raise web.HTTPBadRequest(content_type="application/json", body=err.json())
+        else:
+            return obj
